pyproject.toml
[project]
name = "kadaie1"
version = "0.1.0"
description = "Count short chains in PDB sequence file"
readme = "README.md"
requires-python = ">=3.12"
dependencies = ["biopython"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["kadaie1"]

[project.scripts]
kadaie_1 = "kadaie1.test1:main"